Various code fixes for PR feedback
https://github.com/aws/amazon-vpc-resource-controller-k8s/pull/443
